+++ library/ratchet/src/vscr_ratchet_group_participant.c
@@ -219,7 +219,7 @@ vscr_ratchet_group_participant_cleanup_ctx(vscr_ratchet_group_participant_t *sel
 
 VSCR_PUBLIC void
 vscr_ratchet_group_participant_add_epoch(
-        vscr_ratchet_group_participant_t *self, size_t epoch, vscr_ratchet_chain_key_t **chain_key_ref) {
+        vscr_ratchet_group_participant_t *self, uint32_t epoch, vscr_ratchet_chain_key_t **chain_key_ref) {
 
     VSCR_ASSERT_PTR(self);
 
@@ -250,7 +250,7 @@ vscr_ratchet_group_participant_add_epoch(
 }
 
 VSCR_PUBLIC vscr_ratchet_group_participant_epoch_t *
-vscr_ratchet_group_participant_find_epoch(const vscr_ratchet_group_participant_t *self, size_t epoch) {
+vscr_ratchet_group_participant_find_epoch(const vscr_ratchet_group_participant_t *self, uint32_t epoch) {
 
     VSCR_ASSERT_PTR(self);
