+++ library/ratchet/src/vscr_ratchet_group_ticket.c
@@ -314,7 +314,7 @@ vscr_ratchet_group_ticket_setup_defaults(vscr_ratchet_group_ticket_t *self) {
 
 VSCR_PUBLIC vscr_status_t
 vscr_ratchet_group_ticket_setup_ticket_internal(
-        vscr_ratchet_group_ticket_t *self, size_t epoch, vsc_data_t session_id) {
+        vscr_ratchet_group_ticket_t *self, uint32_t epoch, vsc_data_t session_id) {
 
     VSCR_ASSERT_PTR(self);
 
@@ -357,6 +357,9 @@ err:
     return status;
 }
 
+//
+//  Set session id in case you want to use your own identifier, otherwise - id will be generated for you.
+//
 VSCR_PUBLIC void
 vscr_ratchet_group_ticket_set_session_id(vscr_ratchet_group_ticket_t *self, vsc_data_t session_id) {
 
@@ -388,7 +391,7 @@ vscr_ratchet_group_ticket_generate_key(vscr_ratchet_group_ticket_t *self) {
 }
 
 //
-//  Generates message that should be sent to all participants using secure channel.
+//  Returns message that should be sent to all participants using secure channel.
 //
 VSCR_PUBLIC const vscr_ratchet_group_message_t *
 vscr_ratchet_group_ticket_get_ticket_message(const vscr_ratchet_group_ticket_t *self) {
